+++ CHANGELOG.md
@@ -9,6 +9,58 @@
 > -   :house: [Internal]
 > -   :nail_care: [Polish]
 
+## 3.24.4
+
+#### :boom: Breaking Change
+
+-   Options to hide the functionality of editing directories and files `filebrowser.createNewFolder`, `filebrowser.editImage`,
+    `filebrowser.deleteFolder`,`filebrowser.renameFolder`,`filebrowser.moveFolder`,`filebrowser.moveFile` were marked as deprecated.
+-   Instead added `filebrowser.permissionsPresets: Partial<IPermissions>` option.
+
+Before:
+
+```js
+Jodit.make('#editor', {
+  filebrowser: {
+    createNewFolder: false,
+    deleteFolder: false,
+    renameFolder: false,
+    moveFolder: false,
+    moveFile: false,
+    editImage: false,
+    ajax: {
+      url: 'https://xdsoft.net/jodit/finder/'
+    }
+  },
+});
+```
+
+Now
+```js
+Jodit.make('#editor', {
+  filebrowser: {
+    permissionsPresets: {
+      allowFiles: false,
+      allowFileMove: false,
+      allowFileUpload: false,
+      allowFileUploadRemote: false,
+      allowFileRemove: false,
+      allowFileRename: false,
+      allowFolders: false,
+      allowFolderCreate: false,
+      allowFolderMove: false,
+      allowFolderRemove: false,
+      allowFolderRename: false,
+      allowImageResize: false,
+      allowImageCrop: false,
+    },
+    ajax: {
+      url: 'https://xdsoft.net/jodit/finder/'
+    }
+  },
+});
+```
+
 ## 3.24.3
 
 #### :house: Internal
@@ -42,8 +94,6 @@
  typescript                          ^4.9.3  →     ^4.9.5
  webpack-cli                         ^5.0.0  →     ^5.0.1
 ```
-
-
 
 ## 3.24.2
 
@@ -1926,11 +1976,11 @@ Related with https://github.com/xdan/jodit/issues/574. In some cases need to lim
 -   @property {IUIOption[]} link.selectOptionsClassName=[] The list of the option for the select (to use with
     modeClassName="select")
 -   ex: [
--                                                                                                                                                                                                                                                                                                                                                                                                                                                                   	{ value: "", text: "" },
--                                                                                                                                                                                                                                                                                                                                                                                                                                                                   	{ value: "val1", text: "text1" },
--                                                                                                                                                                                                                                                                                                                                                                                                                                                                   	{ value: "val2", text: "text2" },
--                                                                                                                                                                                                                                                                                                                                                                                                                                                                   	{ value: "val3", text: "text3" }
--                                                                                                                                                                                                                                                                                                                                                                                                                                                                   ]
+-                                                                                                                                                                                                                                                                                                                                                                                                                                                                         	{ value: "", text: "" },
+-                                                                                                                                                                                                                                                                                                                                                                                                                                                                         	{ value: "val1", text: "text1" },
+-                                                                                                                                                                                                                                                                                                                                                                                                                                                                         	{ value: "val2", text: "text2" },
+-                                                                                                                                                                                                                                                                                                                                                                                                                                                                         	{ value: "val3", text: "text3" }
+-                                                                                                                                                                                                                                                                                                                                                                                                                                                                         ]
     PR: https://github.com/xdan/jodit/pull/577 Thanks @s-renier-taonix-fr
 
 ##### New option `statusbar: boolean = true`